pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "fastapi-limiter>=0.1.6",
    "google-cloud-storage>=3.5.0",
    "modal>=1.2.1",
    "pillow>=12.0.0",
    "pinecone>=7.3.0",
    "redis>=5.0.0",
    "transformers>=4.57.1",
]

[dependency-groups]
dev = [
    "notebook>=7.4.7",
]
